refactor(arch): fix missed import paths from domain rename (#209)
- src/commands/ barrel: ../.. → .. (depth 3→2)
- src/commands/show-entry-for-date: ../../model→../model,
  ../../util/paths→../journal/paths,
  ../features/sync-daily-links→../features/sync/sync-daily-links
- src/journal/match-input: ../../util/→../util/, ../../model/→../model/
- src/features/entries/weekly-entry-watcher: paths and sync-daily-links paths
- src/journal/navigation: ../util/paths → ./paths
- src/journal/reader: split util import, add ./paths for resolvePath
- src/ui/codeactions/for-open-tasks: ../commands/ → ../../commands/
- src/test: util/paths → journal/paths

#209
