+++ src/commands/open-next-entry.ts
@@ -10,7 +10,7 @@
 'use strict';
 
 import * as vscode from 'vscode';
-import * as J from '../..';
+import * as J from '..';
 import { AbstractLoadEntryForDateCommand } from './show-entry-for-date';
 import { daysBetween, findAdjacentEntry, getAdjacentWeekInput, resolveAnchor, Mode } from '../journal/navigation';
