+++ src/commands/show-entry-for-input.ts
@@ -18,7 +18,7 @@
 'use strict';
 
 import * as vscode from 'vscode';
-import * as J from '../..';
+import * as J from '..';
 import { AbstractLoadEntryForDateCommand } from './show-entry-for-date';
