+++ src/commands/show-entry-for-tomorrow.ts
@@ -18,7 +18,7 @@
 'use strict';
 
 import * as vscode from 'vscode';
-import * as J from '../..';
+import * as J from '..';
 import { AbstractLoadEntryForDateCommand } from './show-entry-for-date';
