+++ src/commands/show-note.ts
@@ -18,7 +18,7 @@
 'use strict';
 
 import * as vscode from 'vscode';
-import * as J from '../..';
+import * as J from '..';
 
 
 export class ShowNoteCommand implements vscode.Command, vscode.Disposable {
